auth demo

+++ Server/api/module/players/player.service.js
@@ -1,33 +1,62 @@
 const repository = require("./player.repository");
+const authService = require("../auth/auth.service");
 
 const find = async function(query) {
-  return await repository.find(query);
+  let auth = authService.authorization(user, ["admin", "user"]);
+  if (auth) {
+    return await repository.find(query);
+  } else {
+    throw new error("Unauthorized!");
+  }
 };
 
 const findById = async function(id) {
-  return await repository.findById(id);
-};
-const create = async function(data) {
-  if (!data.name || !data.team) {
-    throw new Error("Missing input");
+  let auth = authService.authorization(user, ["admin", "user"]);
+  if (auth) {
+    return await repository.findById(id);
+  } else {
+    throw new error("Unauthorized!");
   }
-  return await repository.create(data);
 };
 
+const create = async function(data) {
+  let auth = authService.authorization(user, ["admin"]);
+  if (auth) {
+    if (!data.name || !data.team) {
+      throw new Error("Missing input");
+    } else {
+      return await repository.create(data);
+    }
+  } else {
+    throw new error("Unauthorized!");
+  }
+};
 
 const update = async function(id, data) {
   const existedData = await repository.findById(id);
-  if (!existedData) {
-    throw new Error("Not found");
+  let auth = authService.authorization(user, ["admin"]);
+  if (auth) {
+    if (!existedData) {
+      throw new Error("Not found");
+    } else {
+      return await repository.update(id, data);
+    }
+  } else {
+    throw new error("Unauthorized!");
   }
-  return await repository.update(id, data);
 };
 const deleteOne = async function(id) {
   const existedData = await repository.findById(id);
-  if (!existedData) {
-    throw new Error("Not found");
+  let auth = authService.authorization(user, ["admin"]);
+  if (auth) {
+    if (!existedData) {
+      throw new Error("Not found");
+    } else {
+      return await repository.delete(id);
+    }
+  } else {
+    throw new error("Unauthorized!");
   }
-  return await repository.delete(id);
 };
 
 module.exports = {
